arc/vet_vso_reloader_populate_join_tables - passed rubocop
